requirements.txt
blinker==1.5
click==8.1.3
Flask==2.1.3
Flask-Mail==0.9.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pyngrok==5.1.0
python-dotenv==0.20.0
PyYAML==6.0
Werkzeug==2.1.2
zipp==3.8.1
